successfully moved SIOC and created actions/reducer logic to use it
