+++ src/components/ApplicationPage/Room.js
@@ -4,20 +4,37 @@ import { connect, } from 'react-redux';
 import io from 'socket.io-client';
 import { withRouter, } from 'react-router';
 import EditorView from './EditorView';
-import { setCreateInput, } from '../../actions/Application';
+import { setCreateInput, deleteLocalUserStream, setCreateVideoFunc, } from '../../actions/Application';
 import { API_BASE_URL, } from '../../config';
+import SIOC from './SIOC';
 import WebCam from './WebCam';
 export const socket = io(API_BASE_URL);
 
 
 export class Room extends React.Component {
+  constructor(props) {
+    super(props);
+    this.SIOC = new SIOC();
+    console.log('Room.js -> new instance of SIOC');
+  }
+
   componentWillMount() {
+    console.log('Room.js -> ComponentDidMount -> init');
+    this.SIOC.init(this.props);
     this.props.dispatch(setCreateInput(this.props.match.params.roomName));
+    this.props.dispatch(setCreateVideoFunc(id => this.SIOC.createVideo(id)));
+
     socket.emit('join room', { room: this.props.match.params.roomName, user: this.props.username, });
+  }
+
+  componentDidMount() {
+    this.SIOC.getLocalUserMedia();
   }
 
   componentWillUnmount() {
     socket.emit('leave room', { room: this.props.match.params.roomName, user: this.props.username, });
+
+    this.props.dispatch(deleteLocalUserStream());
   }
 
   render() {
